pyproject.toml
[project]
name = "android-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "env>=0.1.0",
    "google-genai>=1.10.0",
    "python-dotenv>=1.1.0",
]
